move visitor to tests and clean up a couple of missing returns
I guess this wasn't affecting any test results, but `visit_expr` was missing
early returns after the `Lambda` and `Comp` visits, so it would have been
traversing these nodes twice. I moved the default `walk_expr` call into the
final arm of the match to avoid needing early returns in each other branch.
