+++ crates/ruff_python_parser/src/semantic_errors.rs
@@ -3,7 +3,7 @@
 //! This checker is not responsible for traversing the AST itself. Instead, its
 //! [`SemanticSyntaxChecker::visit_stmt`] and [`SemanticSyntaxChecker::visit_expr`] methods should
 //! be called in a parent `Visitor`'s `visit_stmt` and `visit_expr` methods, respectively.
-use std::{cell::RefCell, fmt::Display};
+use std::fmt::Display;
 
 use ruff_python_ast::{
     self as ast,
@@ -1323,144 +1323,6 @@ pub trait SemanticSyntaxContext {
     fn in_notebook(&self) -> bool;
 
     fn report_semantic_error(&self, error: SemanticSyntaxError);
-}
-
-enum Scope {
-    Module,
-    Function { is_async: bool },
-    Comprehension { is_async: bool },
-}
-
-#[derive(Default)]
-pub struct SemanticSyntaxCheckerVisitor<'a> {
-    checker: SemanticSyntaxChecker,
-    diagnostics: RefCell<Vec<SemanticSyntaxError>>,
-    python_version: PythonVersion,
-    source: &'a str,
-    scopes: Vec<Scope>,
-}
-
-impl<'a> SemanticSyntaxCheckerVisitor<'a> {
-    pub fn new(source: &'a str) -> Self {
-        Self {
-            checker: SemanticSyntaxChecker::new(),
-            diagnostics: RefCell::default(),
-            python_version: PythonVersion::default(),
-            source,
-            scopes: vec![Scope::Module],
-        }
-    }
-
-    #[must_use]
-    pub fn with_python_version(mut self, python_version: PythonVersion) -> Self {
-        self.python_version = python_version;
-        self
-    }
-
-    pub fn into_diagnostics(self) -> Vec<SemanticSyntaxError> {
-        self.diagnostics.into_inner()
-    }
-
-    fn with_semantic_checker(&mut self, f: impl FnOnce(&mut SemanticSyntaxChecker, &Self)) {
-        let mut checker = std::mem::take(&mut self.checker);
-        f(&mut checker, self);
-        self.checker = checker;
-    }
-}
-
-impl SemanticSyntaxContext for SemanticSyntaxCheckerVisitor<'_> {
-    fn seen_docstring_boundary(&self) -> bool {
-        false
-    }
-
-    fn future_annotations_or_stub(&self) -> bool {
-        false
-    }
-
-    fn python_version(&self) -> PythonVersion {
-        self.python_version
-    }
-
-    fn report_semantic_error(&self, error: SemanticSyntaxError) {
-        self.diagnostics.borrow_mut().push(error);
-    }
-
-    fn source(&self) -> &str {
-        self.source
-    }
-
-    fn global(&self, _name: &str) -> Option<TextRange> {
-        None
-    }
-
-    fn in_async_context(&self) -> bool {
-        for scope in &self.scopes {
-            if let Scope::Function { is_async } = scope {
-                return *is_async;
-            }
-        }
-        false
-    }
-
-    fn in_sync_comprehension(&self) -> bool {
-        for scope in &self.scopes {
-            if let Scope::Comprehension { is_async: false } = scope {
-                return true;
-            }
-        }
-        false
-    }
-
-    fn in_module_scope(&self) -> bool {
-        self.scopes
-            .last()
-            .is_some_and(|scope| matches!(scope, Scope::Module))
-    }
-
-    fn in_notebook(&self) -> bool {
-        false
-    }
-}
-
-impl Visitor<'_> for SemanticSyntaxCheckerVisitor<'_> {
-    #[allow(clippy::single_match)]
-    fn visit_stmt(&mut self, stmt: &Stmt) {
-        self.with_semantic_checker(|semantic, context| semantic.visit_stmt(stmt, context));
-        match stmt {
-            Stmt::FunctionDef(ast::StmtFunctionDef { is_async, .. }) => {
-                self.scopes.push(Scope::Function {
-                    is_async: *is_async,
-                });
-                ruff_python_ast::visitor::walk_stmt(self, stmt);
-                self.scopes.pop().unwrap();
-                return;
-            }
-            _ => {}
-        }
-        ruff_python_ast::visitor::walk_stmt(self, stmt);
-    }
-
-    fn visit_expr(&mut self, expr: &Expr) {
-        self.with_semantic_checker(|semantic, context| semantic.visit_expr(expr, context));
-        match expr {
-            Expr::Lambda(_) => {
-                self.scopes.push(Scope::Function { is_async: false });
-                ruff_python_ast::visitor::walk_expr(self, expr);
-                self.scopes.pop().unwrap();
-            }
-            Expr::ListComp(ast::ExprListComp { generators, .. })
-            | Expr::SetComp(ast::ExprSetComp { generators, .. })
-            | Expr::DictComp(ast::ExprDictComp { generators, .. }) => {
-                self.scopes.push(Scope::Comprehension {
-                    is_async: generators.iter().any(|gen| gen.is_async),
-                });
-                ruff_python_ast::visitor::walk_expr(self, expr);
-                self.scopes.pop().unwrap();
-            }
-            _ => {}
-        }
-        ruff_python_ast::visitor::walk_expr(self, expr);
-    }
 }
 
 /// Modified version of [`std::str::EscapeDefault`] that does not escape single or double quotes.
